deployment/requirements.txt
streamlit==1.41.1
sendgrid==6.11
python-dotenv==1.0.1
PyPDF2==3.0.1
openai==1.60.1
langchain==0.3.13
langchain-core==0.3.31
langchain-openai==0.3.2
langchain-chroma==0.1.4
langchain-community==0.3.12
langgraph==0.2.60
opensearch-py==2.8.0
langchain-groq==0.2.1
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.3
apify-client==1.8.1
apify-shared==1.2.1
